fix update
